feature (routing): Native NodeJS Router and RouterHandler interface added.
The aim of this template is to use a few third party libraries as possible. The Router class handles routing by loading each
RouterHandler implementation (via di container) and by looking at the url resource path and request method. RouterHandler is used
as an interface that will help with DIP and OCP. In the di container all possible implementations will be tagged and then used
inside the Router class.